Add namespacing for API
